Fix failing USWDS test.
Why: The exisiting USWDS test checks for a specific semantic version of
USWDS on the 18F website. This will cause this test to fail whenever the
18F website updates it's minor or patch version.

Instead, change this test to look at the major version (in this case v2)
and validate that the field exists.

How: Updating the scanner_ui/api/tests.

Tags: test, bugfix