Add poetry to the CircleCI config.
Why: In order to pin dependencies, replace calls to pip with calls to poetry.
One thing to note: We are creating a requirements.txt file on the fly
using `poetry export -f requirements.txt`. This generates a
requirements.txt with hashes of the

How: Updating the .circleci/config.yml file.

Tags: poetry, circleci