Remove end to end tests for now.
Why: Right now we have a suite of end-to-end tests. These use live page data
[18f.gov, login.gov, afrh.gov, etc] from the Scan Engine. So, while these
tests are a simple way to test the entire system of dependencies, they are
dependent on the data from live pages to operate. This means if the data on
those pages change or if the sites are not available, we get breakages in
our build even if nothing changed on our end. We are also removing the
puppeteer dependencies from the Dockerfile to account for the flakiness with that.

Note: this is a temporary state, and we will add the tests back in with
a new strategy.

Additionally, adding a new endpoint "meta" and one API test for that.

How: Removing the flaky end-to-end tests, adding the meta API endpoint,
updating the Dockerfile.

Tags: testing, Docker, puppeteer.