Remove lighthouse for now. Increase batch size.
Why: Lighthouse scans are consistently running until getting terminated
by the next day's jobs. We are going to turn them off for now.
Additionally, we are not able to run all of the tasks at once on the
cloud.gov instance. This is because the batch size of 3000 generates
nine tasks. Increase to 6000 which should generate five tasks.

How: Changing scan-engine.sh scans, and getdomains.sh batch size.

Tags: bugfix, scan-engine.